+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [ROCK-2116] Handle Discovery Token
 * [QATM-3473] New step to obtain service/container port through Marathon API
 * [QATM-3473] Fix cookies in BaseClient
+* [CON-328] Apply URL encode in URL params when we create secrets through CCT
 
 ## 0.18.0 (August 04, 2020)